refactor: shift company set func to utils
